minor fixes to rewards logic, add import.sql file for initial data load

+++ src/main/java/com/retail/rewards/customerrewardsapi/service/CustomerRewardsService.java
@@ -4,6 +4,6 @@ import java.util.Map;
 
 public interface CustomerRewardsService {
     int getPointsPerMonth(Long customerId, String month);
-    Long getTotalPoints(Long customerId);
+    Integer getTotalPoints(Long customerId);
     Map<String, Object> getMonthlyReport(Long customerId);
 }
